% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alrFeatureSelection.R
\name{alrFeatureSelection}
\alias{alrFeatureSelection}
\title{Additive- Log Ratio (ALR) Feature Selection using BORUTA / LASSO}
\usage{
alrFeatureSelection(
  train_data,
  y_train,
  test_data,
  alr_denom = 1,
  featureSelectionMethod = 1,
  impute_factor = 1e-07,
  num_borutaRuns = 100,
  glm_family = "binomial",
  glm_alpha = 1
)
}
\arguments{
\item{train_data}{a samples by log ratio matrix for feature discovery on training data}

\item{y_train}{training set class labels}

\item{test_data}{a samples by log ratio matrix to subset and return. Note key log-ratios are discovered using training set.}

\item{alr_denom}{which column to use as ALR denominator}

\item{featureSelectionMethod}{1 - Boruta; 2- LASSO ; 3 - None}

\item{impute_factor}{impute factor multiplicative replacement of zeroes}

\item{num_borutaRuns}{number of runs of the Boruta algorithm}

\item{glm_family}{which family to use for logistics regression. Should be a valid glmnet family. default = 'binomial'
#'}

\item{glm_alpha}{glmnet alpha parameter (0,1] where 1-LASSO and (0,1) - elasticnet}
}
\value{
A list containing:\tabular{ll}{
   \code{train_Data} \tab samples by features (derived from featureSelectionMethod) training data  \cr
   \tab \cr
   \code{test_data} \tab samples by features (derived from featureSelectionMethod) test data \cr
   \tab \cr
   }
}
\description{
Additive- Log Ratio (ALR) Feature Selection using BORUTA / LASSO
}
