% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_logratio_network.R
\name{build_logratio_network}
\alias{build_logratio_network}
\title{Build DiCoVar network from logratio coefficients}
\usage{
build_logratio_network(
  mdl,
  s = "lambda.min",
  case_lab = "case",
  ctrl_lab = "control"
)
}
\arguments{
\item{mdl}{Logistic regression model generated by DiCoVar}

\item{s}{Lambda penalty parameter to use when retrieving model coefficients}

\item{case_lab}{Label used for cases when fitting regression model. For
multinomial models, this label MUST match one of the model's class labels.}

\item{ctrl_lab}{Label used for controls when fitting regression model. For
multinomial models, the 'control' group is everything except members of the
'case' group. Best practice would be to label this "other" or similar.}
}
\value{
A list containing:\tabular{ll}{
   \code{keyRats} \tab Data frame with model ratios and coefficients. \cr
   \code{graph} \tab An igraph object containing the logratio network. \cr
   \code{part_weights} \tab Data frame of weights for individual features. \cr
   \tab \cr
}
}
\description{
Takes a set of logratio coefficients from a DiCoVar logistic regression model
and assembles a logratio network.
}
