% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDCV.R
\name{computeDCV}
\alias{computeDCV}
\title{Compute Differential Compositional Variation (DCV)}
\usage{
computeDCV(
  train_data,
  y_train,
  lrs.train = NULL,
  num_repeats = 1,
  num_folds = 5,
  impute_factor = 1e-07,
  rank_orderDCV = F
)
}
\arguments{
\item{train_data}{a set of log-ratio features to be scored}

\item{y_train}{class labels}

\item{lrs.train}{A pre-computed samples by log ratio  matrix}

\item{num_repeats}{number of cross-validation repeats for DCV. Default = 1. Can be computationally expensive to repeat multiple times.}

\item{num_folds}{number of cross-validation folds. Reduces over-fitting.}

\item{impute_factor}{zero-imputation impute factor for multiplicative replacement #'}

\item{rank_orderDCV}{should the log-ratios be order and number of distinct part be computed. Can increase computational time.}
}
\value{
A list containing:\tabular{ll}{
   \code{dcv} \tab Aggregated (across-folds) DCV scores \cr
   \tab \cr
   \code{raw_dcv} \tab raw DCV non aggregated DCV scores with scores from each component \cr
   \tab \cr
}
}
\description{
A function that impute-zeros, closes data and computes DCV for a set of log-ratios. DCV is a univariate 5-component log-ratio scoring algorithm.
Helps to identify important log-ratio feature for classification task.
}
\seealso{
\code{\link[diffCompVarRcpp]{dcvScores}}
}
