% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcv_featureSelection.R
\name{computeDCVStrength}
\alias{computeDCVStrength}
\title{Compute DCV Log Ratio Node Strength}
\usage{
computeDCVStrength(dcv_mat)
}
\arguments{
\item{dcv_mat}{The DCV log ratio matrix. Should be a matrix a p x 2 matrix with columns being the log ratio names and the second column being the DCV score (rowmean).
Usually output from the computeDCV function}
}
\value{
a  numberParts x 2 matrix of node strengths
}
\description{
Calculates the DCV Strength of each node in a log ratio network
}
\seealso{
\code{\link{computeDCV}}
}
