% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcv_featureSelection.R
\name{computeLRs}
\alias{computeLRs}
\title{Impute and Compute all pairwise Log Ratios (PLR)}
\usage{
computeLRs(train_data, impute_factor, y_train)
}
\arguments{
\item{train_data}{count / relative abundance sample by feature matrix where all PLR(p) are to be computed.}

\item{impute_factor}{multiplicative replacement factor to replace zeroes}

\item{y_train}{class labels of data}
}
\value{
a n x p matrix of log ratios
}
\description{
Impute and Compute all pairwise Log Ratios (PLR)
}
