% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcvFeatureSelection.R
\name{dcvFeatureSelection}
\alias{dcvFeatureSelection}
\title{Differential Compositional Variation (DCV) Guided Feature Selection on all Pairwise Log Ratios}
\usage{
dcvFeatureSelection(
  train_data,
  y_train,
  test_data,
  impute_factor = 1e-07,
  dcv_matrix = NULL,
  dcv_method = 2,
  percent_topDCV = 0.95,
  num_sets = 10,
  tarFeats,
  minFeats = 4,
  maxBorutaRuns = 100,
  nruns_rfAUC = 20,
  num_trees = 1000,
  upperBound_percent = 0.5,
  rf_importance = "impurity_corrected"
)
}
\arguments{
\item{train_data}{a samples by log ratio matrix for feature discovery on training data}

\item{y_train}{training set class labels}

\item{test_data}{a samples by log ratio matrix to subset and return. Note key log-ratios are discovered using training set.}

\item{impute_factor}{impute factor multiplicative replacement of zeroes}

\item{dcv_matrix}{Default = NULL; DCV matrix can be supplied}

\item{dcv_method}{1:  Approach-1; 2:  Approach-2; 3:  Approach-3}

\item{percent_topDCV}{Using DCV scores threshold for percent of top log ratios to return (Approach-3)}

\item{num_sets}{number of sets for recursive selection (Approach-2)}

\item{tarFeats}{targeted parts/taxa/etc. to select; (Approach-1)}

\item{minFeats}{Minimum number of log ratio set to consider (Approach-1/2)}

\item{maxBorutaRuns}{number of runs of the Boruta algorithm (Approach-3)}

\item{nruns_rfAUC}{Number of runs for random forest AUC estimate (Approach-2)}

\item{num_trees}{number of trees random forest (Approach-2)}

\item{upperBound_percent}{Max percent of parts/taxa to be considered. Upperbound  = upperBound_percent*total number of parts where sets = seq(upperbound,minFeats,length.out = num_sets) (Approach-2)}

\item{rf_importance}{random forest importance measure (Approach-2)}
}
\value{
A list containing:\tabular{ll}{
   \code{train_Data} \tab samples by features (derived from dcv_method) training data  \cr
   \tab \cr
   \code{test_data} \tab samples by features (derived from dcv_method) test data \cr
   \tab \cr
   }
}
\description{
Performs DCV derived feature selection using three approaches:
\itemize{
 \item{"Approach-1"}{ Targeted featutre selection - (Forward Selection)finds best subset given constraints on the number parts/taxa}
 \item{"DEPRECATED Approach-2"}{ Untargeted network based selection - (Recursive Selection)finds reduced subsets with no constran on numer of parts}
 \item{"DEPRECATED Approach-3"}{ Untargeted DCV guided Boruta - reduces computational time of BORUTA Selection with DCV filtering}
}
}
\details{
@importFrom magrittr %>%
 @importFrom foreach %dopar%#'
}
