% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ridge_coefficients.R
\name{extract_ridge_coefficients}
\alias{extract_ridge_coefficients}
\title{Extract part coefficients from DCV ridge regression model}
\usage{
extract_ridge_coefficients(mdl, s = "lambda.min")
}
\arguments{
\item{mdl}{Logistic regression model generated by DiCoVar}

\item{s}{Lambda penalty parameter to use when retrieving model coefficients}
}
\value{
A tibble containing the coefficients for each part in the model.
}
\description{
Takes a set of logratio coefficients from a DiCoVar logistic regression model
and extracts the coefficients for the individual parts in those logratios.
}
