% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTrainTestSplit.R
\name{extractTrainTestSplit}
\alias{extractTrainTestSplit}
\title{Extract Train-Test Split}
\usage{
extractTrainTestSplit(
  foldDataList,
  fold,
  maxSparisty = 0.9,
  permLabels = F,
  extractTelAbunance = T
)
}
\arguments{
\item{foldDataList}{partition fold data. Usually output from the partition data function}

\item{fold}{which fold to extract}

\item{maxSparisty}{percentage of samples with 0 counts for a each feature/column/taxa/etc.}

\item{permLabels}{should the train data labels be permuted}

\item{extractTelAbunance}{should data be closed and imputed (bool)? If not, the is returned as-is in the train-test split.}
}
\value{
A list containing:\tabular{ll}{
   \code{train_Data} \tab samples by features training data  \cr
   \tab \cr
   \code{test_data} \tab samples by features test data \cr
   \tab \cr
   \code{train_ids} \tab row metadata for training samples  \cr
   \tab \cr
   \code{test_ids} \tab row metadata for test samples  \cr
   \tab \cr
   \code{y_train} \tab training class/group labels  \cr
   \tab \cr
   \code{y_test} \tab test class/group labels \cr
   \tab \cr
   \code{minCLass} \tab minority class for AUPRC calculations and Class imbalance Problems. *Valid for binary class labels only*  \cr
   \tab \cr
   \code{majClass} \tab minority class for AUPRC calculations and Class imbalance Problems. *Valid for binary class labels only*   \cr
   \tab \cr
}
}
\description{
Function to extract partitioned train-test. Extract row metadata, performs zero-imputation and returns labels for train-test data
}
