% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{getLogratioFromList}
\alias{getLogratioFromList}
\title{Get log-ratios from a list of ratio names}
\usage{
getLogratioFromList(Ratio, raMatrix, Class)
}
\arguments{
\item{Ratio}{Ratio names where numerator and denominator are separated by '___'. Numerator and Denominator names should match column names of raMatrix}

\item{raMatrix}{a zero-imputed relative abundance or count matrix where the log-ratios will be computed from}

\item{Class}{class labels for raMAtrix samples.}
}
\value{
a set of logratios corresponding the the Ratio list input provided computed from the raMatrix
}
\description{
Get log-ratios from a list of ratio names
}
