% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcv_featureSelection.R
\name{hybrid_dcvfeatureSelection}
\alias{hybrid_dcvfeatureSelection}
\title{Performs DCV Hybrid Feature Selection via Hub Node Significance}
\usage{
hybrid_dcvfeatureSelection(
  xtrain,
  ytrain,
  xtest = NULL,
  impute_factor = 1e-07,
  useGLM = F,
  nperms = 20,
  nfold_dcv = 1,
  alpha = 0.1,
  num_treesRFE = 2000,
  num_setRFE = 20,
  verbose = 0,
  rfImportance = "impurity",
  k_fold = 10,
  minPercFeat = 0.2,
  maxBorutaRuns = 200
)
}
\arguments{
\item{xtrain}{A samples(n) by log ratios(p) matrix to be used for feature selection}

\item{ytrain}{Class labels vector of length n}

\item{xtest}{A samples (xn)by log ratios (xp) matrix to subset after features are discovered on the xtrain matrix.}

\item{impute_factor}{multiplicative factor for imputed zero counts/abundance}

\item{useGLM}{should penalized regression be used instead of hybrid}

\item{nperms}{number of permutation for hub node discovery}

\item{nfold_dcv}{number of cross validation folds for DCV computation. Default is 1(all data). Increasing folds increases computational time exponentially.}

\item{alpha}{significance threshold for hub node identification. Default is 0.10. Lower alpha returns fewer hubs and vice-versa.}

\item{num_treesRFE}{number of Random Forest trees to be grown during recursive feature elimination (RFE). Default = 2,000}

\item{num_setRFE}{number of sets in RFE. Default = 20}

\item{verbose}{should status of RFE and Boruta be displayed}

\item{rfImportance}{RF feature importance metric to be used with RFE. Requires a valid 'ranger::ranger' importance measure}

\item{k_fold}{number of cross-validation folds for RFE}

\item{minPercFeat}{minimum number of feature to be considered in RFE}

\item{maxBorutaRuns}{max number of Boruta runs}
}
\value{
A list containing:\tabular{ll}{
   \code{MST} \tab a n x f (retained log ratios) derived from MST pruned stage-1 log ratio network \cr
   \tab \cr
   \code{Dense} \tab a n x f (retained log ratios)  derived from dense stage-1 network  \cr
   \tab \cr
   }
}
\description{
Selects 2 subset of log ratios containing statistically significant hub nodes (permutation testing). Subsets returned are first in three stages.
In stage 1 log-ratios are scored using DCV. Afterwards hub 'parts/taxa/etc.' are identified within the log ratio network and significant (permutation thresholded) log ratios containing
at least one hub are retained. Finally in stages 2 & 3 two rounds of feature selection is carried out the final subsets. Theses subset are:
\itemize{
 \item{"MST-set"}{Non Redundant Subset of logratios pruned after stage 1}
 \item{"Dense"}{ All log ratios from stage 1 }
}
}
