% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{knnADJtoSYM}
\alias{knnADJtoSYM}
\title{Make a KNN or non-symmetric adjacency matrix symmetric}
\usage{
knnADJtoSYM(knnADJ_MAT)
}
\arguments{
\item{knnADJ_MAT}{a non-symmetric adjacency matrix to be made symmetric}
}
\value{
a symmetric adjacency matrix
}
\description{
Make a KNN or non-symmetric adjacency matrix symmetric
}
