% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{lodo_partition}
\alias{lodo_partition}
\title{Leave-One-Dataset-Out (LODO) cross validation partitioning}
\usage{
lodo_partition(df, dataset_labels, seed = 8272008)
}
\arguments{
\item{df}{samples by features matrix to be partitioned. First columns must contain class labels.}

\item{dataset_labels}{vector of dataset label for each sample in df}

\item{seed}{random seed}
}
\value{
a k-element list containing the train-test (xtrain_combinedFolds / xtest_kthFold) splits along with sample IDs (xtrain_IDs / xtest_IDs )#' @export
}
\description{
Leave-One-Dataset-Out (LODO) cross validation partitioning
}
