% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcv_featureSelection.R
\name{rfeSelection.ByMetric}
\alias{rfeSelection.ByMetric}
\title{Recursive Feature Elimination (RFE) using random forest feature importance}
\usage{
rfeSelection.ByMetric(
  train_ratio,
  test_ratio,
  ytrain,
  metric = "multiClass.ROC",
  kfold = 10,
  impMeasure = "impurity_corrected",
  sets = 10,
  ntrees = 2000,
  minPercentFeatReturn = 0.2
)
}
\arguments{
\item{train_ratio}{training log ratios for feature subset discovery. A sample by log ratio matrix}

\item{test_ratio}{test log ratios to be subset after feature discovery on training dats}

\item{ytrain}{training class labels}

\item{metric}{OOB error metric. Currently c("multiClass.ROC"(default),"AUPRC","AUROC")}

\item{kfold}{number of folds of nested cross validation}

\item{impMeasure}{RF Feature importance metric.}

\item{sets}{number of recursive set to consider}

\item{ntrees}{number of RF trees to grow for each set}

\item{minPercentFeatReturn}{min percent of log ratio features to return}
}
\value{
A list containing:\tabular{ll}{
   \code{reducedTrainRatios} \tab a n x f (retained log ratios)  matrix with RFE selected log ratios \cr
   \tab \cr
   \code{reducedTestRatio} \tab a n x f (retained log ratios)  matrix with train set discovered log ratios \cr
   \tab \cr
   \code{trainPerformance} \tab RFE performance results  \cr
   \tab \cr
   }
}
\description{
Performs nested RFE using a OOB AUC of a random forest (RF) model. Features are ranked using common RF feature importance.
RFE is implemented using the ranger R package where feature importance measures should correspond to valid ranger metrics.
}
\seealso{
\code{\link[ranger]{ranger}}
}
