% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{rocPlot}
\alias{rocPlot}
\title{A pROC derived wrapper for computing sens-spec threshold data fro plotting ROC curves and confidence  intervals}
\usage{
rocPlot(mroc)
}
\arguments{
\item{mroc}{a multi-class ROC object from pROC::multiclass.roc()}
}
\value{
a data.frame with sens., specificity and confidence interval
}
\description{
Provides an easy wrapper for extract ROC curve data from
}
\seealso{
\code{\link[pROC]{multiclass.roc}}
}
