% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeted_dcvSelection.R
\name{targeted_dcvSelection}
\alias{targeted_dcvSelection}
\title{Targeted DCV Feature Selection}
\usage{
targeted_dcvSelection(
  trainx,
  testx,
  y_label,
  y_test,
  dcv = NULL,
  lrs.train = NULL,
  lrs.test = NULL,
  minConnected = F,
  ensemble = c("ranger", "pls", "svmRadial", "glmnet", "rangerE"),
  tarFeatures = 5,
  imp_factor = 1e-07,
  select_randomFeatures = F,
  use_rfe = T,
  alpha_ = 0,
  ts.id,
  seed = 8272008,
  max_sparsity = 0.9,
  useRidgeWeights = T,
  scaledata = T
)
}
\arguments{
\item{trainx}{train data partition samples x parts/taxa/etc.}

\item{testx}{test data partition samples x parts/taxa/etc.}

\item{y_label}{train labels}

\item{y_test}{test labels}

\item{dcv}{a dcv matrix can optionally be provided}

\item{lrs.train}{An optional supplied logratio matrix for training data; must be supplied if dcv scores are pre computed}

\item{lrs.test}{An optional supplied logratio matrix for test data; must be supplied if dcv scores are pre computed}

\item{minConnected}{should the ratio network be minimally connected or threshold with dcv_score>0}

\item{ensemble}{ensemble model definition, or FALSE to skip fitting ensemble models}

\item{tarFeatures}{number of parts/taxa/etc. to retain (independent from number of ratios to retain)}

\item{imp_factor}{factor to multiplicative impute count zeros}

\item{select_randomFeatures}{should random parts/taxa/etc. be selected; useful for benchmarking}

\item{use_rfe}{should a random forest recursive elimination model be trained}

\item{alpha_}{glmnet alpha parameter for fitting the final ridge regression model}

\item{ts.id}{test set id matrix. Can be found in output list from partitioning functions i.e.  kfoldDataPartition(),lodo_partition(),etc. Can be manually entered nrows = nrows(testx)}

\item{seed}{Random Seed control for reproducibility}

\item{max_sparsity}{max sparsity of parts/taxa/etc. For example max_sparsity=0.10 would mean to only retain parts/taxa/etc present in at least 10\% of all samples.}

\item{useRidgeWeights}{should ensemble model first be weighted by ridge regression coefficients?}

\item{scaledata}{should train data be scaled (test data scaled based on train)}
}
\value{
a list containing the models and diagnostic outputs
}
\description{
Performs DCV feature selection to build a model with a specific number of parts.
}
