% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.dicovar.R
\name{unsupervised.dicovar}
\alias{unsupervised.dicovar}
\title{Unsupervised Differential Compositional Variation (DiCoVar) Analysis}
\usage{
unsupervised.dicovar(
  X,
  max_sparsity = 0.9,
  replace = TRUE,
  dcv_folds = 1,
  dcv_repeats = 1,
  scale_data = TRUE,
  alpha = 0,
  seed = NULL
)
}
\arguments{
\item{X}{matrix-like object with one row per sample and one column per part}

\item{max_sparsity}{Parts with a higher proportion of zero/missing values will be dropped before zero imputation.}

\item{replace}{Should fake data be sampled from real data with replacement?}

\item{dcv_folds}{Partition the data into folds and average DCV scores from each fold}

\item{dcv_repeats}{How many times should DCV repeat the partitioning process?}

\item{scale_data}{Should data be scaled to unit standard deviation before ridge regression?}

\item{alpha}{amount of LASSO regression to mix into the final ridge regression model}

\item{seed}{random seed to use for reproducible results, or NULL to use the current seed}
}
\value{
A list containing:\tabular{ll}{
   \code{Performance} \tab How well a glmnet model can distinguish real from fake data. \cr
   \code{glm_model} \tab The glmnet model itself. \cr
   \code{part_matrix} \tab The closed, zero-imputed part matrix. Includes fake data. \cr
   \code{dcv_scores} \tab The DCV score data for all logratios. \cr
   \code{ridge_pmat} \tab Fitted class probabilities from the glmnet model. \cr
   \tab \cr
}
}
\description{
A function that takes raw data and constructs a synthetic copy by resampling
each part. These data sets are transformed into the logratio space, and
DCV scores used to select logratios which differ in the real data versus the
resampled data.
}
\seealso{
\code{\link[diffCompVarRcpp]{dcvScores}}
}
