% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcv_featureSelection.R
\name{dcvRatioFilter}
\alias{dcvRatioFilter}
\title{Performs DCV Ratio Filtering}
\usage{
dcvRatioFilter(
  xtrain,
  ytrain,
  xtest = NULL,
  impute_factor = 1e-07,
  useGLM = F,
  th_percent = 0.5
)
}
\arguments{
\item{xtrain}{A samples(n) by log ratios(p) matrix to be used for feature selection}

\item{ytrain}{Class labels vector of length n}

\item{xtest}{A samples (xn)by log ratios (xp) matrix to subset after features are discovered on the xtrain matrix.}

\item{impute_factor}{multiplicative factor for imputed zero counts/abundance}

\item{useGLM}{should penalized regression be used instead of hybrid}

\item{th_percent}{dcv score percentile for thresholding i.e. keep ratio where dcvScore > threshold(th_percent)}
}
\value{
A list containing:\tabular{ll}{
   \code{MST} \tab a n x f (retained log ratios) derived from MST  \cr
   \tab \cr
   \code{Dense} \tab a n x f (retained log ratios)  derived from dense thresholding \cr
   \tab \cr
   }
}
\description{
Selects 2 subset of log ratios based dcvScore thresholding. Theses subset are:
\itemize{
 \item{"MST-set"}{ - Non Redundant Subset of logratios pruned after thresholding}
 \item{"Dense"}{ - All log ratios after thresholding }
}
}
