% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativeAbundanceFeatureSlection.R
\name{relAbundanceFeatureSelection}
\alias{relAbundanceFeatureSelection}
\title{Relative Abundance Feature Selection}
\usage{
relAbundanceFeatureSelection(
  train_data,
  y_train,
  test_data,
  impute_zeroes = T,
  featureSelectionMethod = 1,
  impute_factor,
  num_borutaRuns = 100,
  glmRepeats = 5,
  glmFolds = 5
)
}
\arguments{
\item{train_data}{a samples by log ratio matrix for feature discovery on training data}

\item{y_train}{training set class labels}

\item{test_data}{a samples by log ratio matrix to subset and return. Note key log-ratios are discovered using training set.}

\item{impute_zeroes}{should zeroes be imputed}

\item{featureSelectionMethod}{1 - Boruta; 2- LASSO ; 3 - None}

\item{impute_factor}{impute factor multiplicative replacement of zeroes}

\item{num_borutaRuns}{number of runs of the boruta algorithm}

\item{glmRepeats}{number of cv repeats}

\item{glmFolds}{number of cvFolds}
}
\value{
A list containing:\tabular{ll}{
   \code{train_Data} \tab samples by features (derived from featureSelectionMethod) training data  \cr
   \tab \cr
   \code{test_data} \tab samples by features (derived from featureSelectionMethod) test data \cr
   \tab \cr
   }
}
\description{
*Not recommended.* Carries out Boruta feature selection on raw count of relative abundance data.
}
